#126 Refactoring OpenActionFragment by calling a new Activity from it
- Created activity WorkoutReschedulingScreen that is created every time
a user clicks on "Add to workout mode" in OpenActionFragment
- Added to ActivityComponent for dagger injection
- Added to manifest.xml
- Removed WorkoutLauncherDialogFragment class, not needed anymore
- In OpenActionFragment, removed the code that was created in the
workoutMode selection.
- Moved that code to WorkoutReschedulingScreen
- Refactored the logic of setting the cards date and user error
handling as follow:
1) The user error handling is divided into date validation
(whether the user chose a date) and number of days / number of cards
validation (whether the user chose either a day or cards, both of which cannot be equal to 0, empty, or larger than the number of cards in the deck
2) Once the user presses on add to workout mode, in the logic of the onClickListener for the OK button,  we verify that the error
handling returns true (meaning that no error are made on the part of the
user, inputs are all there and correct)
3) In the onClickListener for the OK butotn, we then retrieve the date and the number of cards / days the user
entered and schedule the cards for a new workout.